pyproject.toml
[project]
name = "gofreddy"
version = "0.1.0"
description = "Distribution engineering agency CLI — AI-powered distribution systems for clients"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "google-genai>=1.0.0",
    "google-adk==1.27.2",
    "yt-dlp>=2024.0.0",
    "scikit-learn>=1.4.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "apify-client>=1.0.0",
    "requests>=2.28.0",
    "aioboto3>=13.0.0",
    "aiofiles>=24.0.0",
    "httpx>=0.27.0",
    "cachetools>=5.0.0",
    "orjson>=3.0.0",
    "tenacity>=8.0.0",
    "aiolimiter>=1.1.0",
    "xpoz>=0.2.0",
    "opentelemetry-instrumentation-google-genai>=0.1.0",
    "mistune>=3.0.0",
    "fal-client>=0.13.1,<1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "gpyrobotstxt>=0.1.0",
    "dataforseo-client>=0.2.0",
    "weasyprint>=62.0",
    "markdown>=3.5.0",
    "nh3>=0.2.14",
    "pymupdf>=1.24.0",
    "feedparser>=6.0.0",
    "fish-audio-sdk>=1.0.0",
    "fpdf2>=2.7.0",
    "openai>=1.80.0",
    "email-validator>=2.2.0",
    "xai-sdk>=1.8.0,<2.0.0",
    "PyYAML>=6.0",
    "typer>=0.12.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
music = ["librosa>=0.10"]
gsc = [
    "google-auth>=2.0.0",
    "google-api-python-client>=2.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "moto[s3]>=5.0.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
    "respx>=0.21.0",
    "freezegun>=1.2.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-aiofiles>=24.0.0",
    "types-cachetools>=5.0.0",
]

[project.scripts]
freddy = "cli.freddy.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "cli"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
addopts = []
markers = [
    "gemini: uses real Gemini API",
    "live_api: exercises live app/services over real API boundaries",
    "r2: uses real R2 storage",
    "external_api: calls ScrapeCreators/Apify/yt-dlp",
    "deepfake: uses LIPINC/Reality Defender",
    "mock_required: genuinely needs mocks (cannot trigger real error)",
    "slow: takes >10 seconds",
    "spike: validation spike (requires real Gemini + curated videos)",
]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
disable_error_code = ["call-arg", "arg-type", "union-attr", "assignment", "has-type"]
exclude = ["tests/"]

[tool.ruff]
target-version = "py313"
exclude = []
